make switch button disable component
